2022/requirements.txt
pandas
requests
beautifulsoup4
nltk
lxml
scikit-learn
streamlit
firecrawl-py
python-dotenv
